ISIS-3207: wicket viewer (ui); move test classes into /test subpack
