--- viewers/wicket/ui/src/test/java/org/apache/isis/viewer/wicket/ui/ComponentType_ConstantsTest.java
@@ -1,38 +1,0 @@
-/*
- *  Licensed to the Apache Software Foundation (ASF) under one
- *  or more contributor license agreements.  See the NOTICE file
- *  distributed with this work for additional information
- *  regarding copyright ownership.  The ASF licenses this file
- *  to you under the Apache License, Version 2.0 (the
- *  "License"); you may not use this file except in compliance
- *  with the License.  You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package org.apache.isis.viewer.wicket.ui;
-
-import org.junit.Test;
-
-import org.apache.isis.viewer.commons.model.components.UiComponentType;
-
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-public class ComponentType_ConstantsTest {
-
-    @Test
-    public void testToString() {
-        assertThat(UiComponentType.COLLECTION_CONTENTS.toString(), is("collectionContents"));
-    }
-
-    public enum X {
-        FOO, BAR
-    }
-}
